# backend/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: PBCEx API
  version: 1.0.0
  description: |
    People's Bank & Commodities Exchange (PBCEx) API for precious metals trading, 
    custody, and fulfillment operations.

    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```

    ## Rate Limits
    - General API: 1000 requests per 15 minutes
    - Authentication: 5 attempts per 15 minutes  
    - KYC submissions: 10 per hour
    - Trading: 30 per minute

  contact:
    name: PBCEx API Support
    email: api-support@pbcex.com
  license:
    name: Proprietary
  x-build-info:
    branch: feature/money-movement-backend
    shortHash: ff3d03a
    generatedAt: '2025-08-31T20:00:00Z'
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.pbcex.com
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          enum:
            [
              VALIDATION_ERROR,
              AUTHENTICATION_ERROR,
              AUTHORIZATION_ERROR,
              NOT_FOUND,
              CONFLICT,
              RATE_LIMITED,
              INTERNAL_ERROR,
              SERVICE_UNAVAILABLE,
            ]
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        details:
          type: object

    # Money Movement Schemas
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/qr/pay:
    post:
      tags: [QR]
      operationId: createPayToken
      summary: Create QR payment token
      description: Generate a QR token for receiving payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset, amount, expiresIn]
              properties:
                asset:
                  type: string
                  enum: [AU, AG, PT, PD, CU, USD, USDC]
                  description: Asset to receive
                amount:
                  type: string
                  pattern: '^\d+\.?\d*$'
                  description: Amount to receive
                expiresIn:
                  type: integer
                  minimum: 300
                  maximum: 86400
                  description: Token expiration in seconds (5 min to 24 hours)
                memo:
                  type: string
                  description: Optional payment memo
      responses:
        '201':
          description: QR payment token created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: QR token identifier
                      qrCode:
                        type: string
                        description: Base64 encoded QR code image
                      expiresAt:
                        type: string
                        format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-feature-flag: MONEY_MOVEMENT_ENABLED
  /api/qr/receive:
    post:
      tags: [QR]
      operationId: createReceiveToken
      summary: Create QR receive token
      description: Generate a QR token for sending payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset, amount, expiresIn]
              properties:
                asset:
                  type: string
                  enum: [AU, AG, PT, PD, CU, USD, USDC]
                  description: Asset to send
                amount:
                  type: string
                  pattern: '^\d+\.?\d*$'
                  description: Amount to send
                expiresIn:
                  type: integer
                  minimum: 300
                  maximum: 86400
                  description: Token expiration in seconds
                memo:
                  type: string
                  description: Optional payment memo
      responses:
        '201':
          description: QR receive token created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: QR token identifier
                      qrCode:
                        type: string
                        description: Base64 encoded QR code image
                      expiresAt:
                        type: string
                        format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-feature-flag: MONEY_MOVEMENT_ENABLED
  /api/qr/{token}:
    get:
      tags: [QR]
      operationId: getQRToken
      summary: Get QR token details
      description: Retrieve QR token information (public access)
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: QR token identifier
      responses:
        '200':
          description: QR token details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    $ref: '#/components/schemas/QRToken'
        '404':
          description: QR token not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/qr/{token}/use:
    post:
      tags: [QR]
      operationId: useQRToken
      summary: Use QR token
      description: Consume a QR token to complete payment
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: QR token identifier
      responses:
        '200':
          description: QR token used successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Payment completed successfully'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      x-feature-flag: MONEY_MOVEMENT_ENABLED
  /api/qr/{token}/cancel:
    post:
      tags: [QR]
      operationId: cancelQRToken
      summary: Cancel QR token
      description: Cancel an active QR token
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: QR token identifier
      responses:
        '200':
          description: QR token cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'QR token cancelled successfully'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      x-feature-flag: MONEY_MOVEMENT_ENABLED
  /api/qr/history:
    get:
      tags: [QR]
      operationId: getQRHistory
      summary: Get QR token history
      description: Retrieve user's QR token usage history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: QR history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/QRToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-feature-flag: MONEY_MOVEMENT_ENABLED
  /api/qr/stats:
    get:
      tags: [QR]
      operationId: getQRStats
      summary: Get QR usage statistics
      description: Retrieve QR token usage statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: QR stats retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'SUCCESS'
                  data:
                    type: object
                    properties:
                      totalTokens:
                        type: integer
                      activeTokens:
                        type: integer
                      completedPayments:
                        type: integer
                      totalVolume:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-feature-flag: MONEY_MOVEMENT_ENABLED

  # Payment Requests Endpoints
...